backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
markdown
coverage
pytest
pytest-django
